refactor: switch page between announce and classroom
